requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
Django==2.2.14
django-cors-headers==3.5.0
django-extra-views==0.11.0
django-haystack==2.8.1
django-oscar==2.0.4
django-oscar-api==2.0.2
-e git+git@github.com:hurta2yaisel/django-oscar-api-vue-storefront.git@07714c5278e5a54ef3f8261597e60526050a5550#egg=django_oscar_api_vue_storefront
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
elasticsearch==7.9.1
elasticsearch-dsl==7.3.0
factory-boy==2.12.0
Faker==4.14.0
idna==2.10
phonenumbers==8.12.11
Pillow==7.2.0
purl==1.5
pycountry==20.7.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.25.11
